recsys/requirements.txt
pandas==1.5.0
pydantic==1.10.2
tqdm==4.64.1
scikit-surprise==1.1.3
requests==2.28.1
